feat: Add sample-requests.http for testing API locally (#36)
Add sample-requests.http. This allows you to run the system locally and test HTTP requests to easily verify that the endpoints work as expected. You can run .http files in IntelliJ IDEA and it will display whether the requests succeeded or failed. This is a faster alternative than opening the Swagger UI to test the endpoints.